requirements.txt
django==2.2.2
django-annoying==0.10.4
django-autocomplete-light==3.3.5
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-extensions==2.1.9
django-filter==2.1.0
django-jet==1.0.8
django-picklefield ==2.0
django-registration-redux==2.6
django-stronghold==0.3.0
django-widget-tweaks==1.4.5
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
mysqlclient==1.4.2.post1
redis==3.2.1
requests==2.22.0
xlrd==1.2.0
xlwt==1.3.0
